made constants and handler packages internal
