+++ addons/binding/org.openhab.binding.fronius/src/main/java/org/openhab/binding/fronius/internal/FroniusBridgeConfiguration.java
@@ -6,15 +6,15 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.fronius;
+package org.openhab.binding.fronius.internal;
 
 /**
- * The {@link FroniusBaseDeviceConfiguration} is the class used to match the
- * thing configuration.
+ * The {@link FroniusBridgeConfiguration} is the class used to match the
+ * bridge configuration.
  *
  * @author Thomas Rokohl - Initial contribution
  */
-public class FroniusBaseDeviceConfiguration {
-    public Integer deviceId;
-
+public class FroniusBridgeConfiguration {
+    public String hostname;
+    public Integer refreshInterval;
 }
